TWRP-ify
For some reason this tree does not generate the dts folder for the
dt.img so we are using a prebuilt dt.img. Use a prebuilt init
binary as the CM init binary does not contain all the things
needed to generate the by-name devices. Add twrp.fstab.

A few ugly hacks in the mkbootimg.mk but I do not have the device
so cleaning these up is a bit hard. It works and that is good
enough for now.

Conflicts:
	device.mk